+++ packages/core/src/defaults/values/effect-values.ts
@@ -18,7 +18,7 @@ import {
 } from "../variables";
 
 export const blurValues: Record<string, string> = {
-	none: `blur(${blur * 0})`,
+	none: `blur()`,
 	xs: `blur(${blur}px)`,
 	sm: `blur(${blur * 2}px)`,
 	md: `blur(${blur * 4}px)`,