+++ packages/core/src/defaults/variables.ts
@@ -78,11 +78,11 @@ export const topRightBottomLeft = 0.25;
 
 // font
 export const fontMono =
-	"ui-monospace, 'Cascadia Code', 'Source Code Pro', Menlo, 'Roboto Mono', 'Ubuntu Mono', Consolas, 'Liberation Mono', 'Courier New', monospace";
+	"system-ui, -apple-system, Segoe UI, Roboto, Helvetica Neue, Noto Sans, Liberation Sans, Arial, sans-serif, Apple Color Emoji, Segoe UI Emoji, Segoe UI Symbol, Noto Color Emoji";
 export const fontSerif =
-	"'Iowan Old Style', 'Palatino Linotype', 'URW Palladio L', P052, serif";
+	"ui-monospace, Cascadia Code, Source Code Pro, Menlo, Roboto Mono, Ubuntu Mono, Consolas, Liberation Mono, Courier New, monospace";
 export const fontDefault =
-	"system-ui, -apple-system, 'Segoe UI', Roboto, 'Helvetica Neue', 'Noto Sans', 'Liberation Sans', Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji'";
+	"Iowan Old Style, Palatino Linotype, URW Palladio L, P052, serif";
 
 export const fontSizeXs = 0.75;
 export const fontSizeSm = 0.875;