+++ packages/core/src/defaults/variants/media-queries.ts
@@ -1,8 +1,8 @@
 export const mediaQueries = [
 	{ prefix: "pc", value: "@media (pointer: coarse)" },
-	{ prefix: "sm", value: "@media (width >= 40rem)" },
-	{ prefix: "md", value: "@media (width >= 48rem)" },
-	{ prefix: "lg", value: "@media (width >= 64rem)" },
-	{ prefix: "xl", value: "@media (width >= 80rem)" },
-	{ prefix: "xxl", value: "@media (width >= 96rem)" },
+	{ prefix: "sm", value: "@media (min-width: 40rem)" },
+	{ prefix: "md", value: "@media (min-width: 48rem)" },
+	{ prefix: "lg", value: "@media (min-width: 64rem)" },
+	{ prefix: "xl", value: "@media (min-width: 80rem)" },
+	{ prefix: "xxl", value: "@media (min-width: 96rem)" },
 ] as const;